URLDispatcher configuration improvements
The list of URLs to use for different browsers is now read from a JSON
file, which prevents having to update the configuration every time a URL
is added, and makes it possible to keep it private. The file is re-read
automatically when it changes, so it's not necessary anymore to restart
Hammerspoon to add/remove URLs.